commit-2 : IPFS-pinata part complete and clean code
